Issue #8: Make a recharge offer to a customer
Wrote a php script to make a recharge offer.  Point to be remembered is that
money is always stored as paise in the database.  And it will always be used as
paise in the program.  Only during display to end user we need to display in
decimal (rupee.paise) format. Also, the offers are to be made in terms of
cashback and not discount percentage.  I think cashback is a better way to
provide incentive.